+++ bin/video/video_generator.dart
@@ -112,24 +112,24 @@ class VideoGenerator implements Generator {
       args.add(outputPath);
 
       log.i('Encoding: $filename');
+
       final ProcessResult result = await Process.run('ffmpeg', args);
 
-      final File file = File(outputPath);
-
-      // Check if the encoding was successful and the output file is valid.
-      if (result.exitCode != 0 ||
-          !file.existsSync() ||
-          file.lengthSync() == 0) {
-        log.e('Error encoding $filename: ${result.stderr}');
-        // Clean up any invalid output file.
-        if (file.existsSync()) {
-          file.deleteSync();
-        }
+      if (result.exitCode != 0) {
+        throw EncodingException.fromResult(filename, result);
       }
+    } on EncodingException catch (e) {
+      log.e(e.message);
     } on UnsupportedException catch (e) {
       log.w(e.message);
     } catch (e) {
       log.e('Exception encoding $filename: $e');
+    } finally {
+      final File file = File(outputPath);
+      if (file.existsSync() && file.lengthSync() == 0) {
+        log.w('Cleaning up invalid output file: $filename');
+        file.deleteSync();
+      }
     }
   }
